requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.12.3
pydantic[email]==2.12.3
requests==2.32.5
##schedule==1.2.2
python-multipart==0.0.9

# Weather and email dependencies
requests==2.32.5

# Environment variables
python-dotenv==1.0.0